fix: Resolve API errors and enable development mode
- Fix duplicate /api prefix in all API endpoints (Dashboard, Expenses, Goals, AddTransactionForm, accountsApi, incomeSourcesApi)
- Update Vite proxy from port 8081 to 8080
- Switch database from PostgreSQL to H2 in-memory for development
- Fix type mismatch in FinancialSummaryResponse (double to BigDecimal, int to Long)
- Add BigDecimal import to TransactionService
- Bypass authentication in development mode (ProtectedRoute)
- Add AuthCallbackPage for OAuth2 flow
- Update authStore and main.tsx for better auth handling
- Backend now runs successfully on port 8080 with H2 database
- Frontend runs on port 3000 with working API proxy

+++ services/finance-service/src/main/java/com/aifa/finance/dto/FinancialSummaryResponse.java
@@ -1,18 +1,19 @@
 package com.aifa.finance.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
+import lombok.Builder;
 import lombok.NoArgsConstructor;
+import lombok.AllArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class FinancialSummaryResponse {
-    private Double totalIncome;
-    private Double totalExpenses;
-    private Double netSavings;
-    private Integer transactionCount;
+    private BigDecimal totalIncome;
+    private BigDecimal totalExpenses;
+    private BigDecimal netSavings;
+    private Long transactionCount;
     private String period;
 }
